refactor: restrucutre bookmarks component inside a dir
