+++ src/components/Bookmarks/index.ts
@@ -1,0 +1,3 @@
+export { default as BookmarksList } from './List'
+export { default as BookmarksItem } from './Item'
+export * from './types'
