+++ src/components/Bookmarks/utils.ts
@@ -1,0 +1,3 @@
+import { Folder } from './types'
+
+export const checkIsFolder = ( item: Folder ): item is Folder => !!( item ).bookmarks
